==> 09-doing-gis-in-geopandas.ipynb <==
%matplotlib inline
import pandas as pd
import geopandas as gpd
import matplotlib.pyplot as plt

# Doing GIS in `geopandas`
`geopandas` brings fundamental GIS operations to the `pandas` environment, allowing you to construct data pipelines in code that can automate routing geospatial data processing workflows. For the purposes of the quick survey of methods in this session I'll split operations into _unary_ operations on a single datasets and _binary_ operations between two datasets. Typical unary operations include buffering and finding representative points. You can think of these as most often being useful in transforming geospatial data from one form to another. Binary operations allow you to join spatial datasets together based on spatial relations between them and include overlay, clipping, and filtering.

## Data
For these notebooks, I grabbed data from [this web page](https://opendata.gw.govt.nz/maps/GWRC::metlink-gtfs/about) relating to 'stops' and 'routes' in Wellington, along with our old friend the SA2 polygons.

stops = gpd.read_file("data/metlink-stops.gpkg").to_crs(2193)

routes = gpd.read_file("data/metlink-routes.gpkg").to_crs(2193)

sa2 = gpd.read_file("data/sa2-wellington.gpkg")

We can see what we are dealing with using a web map.

m = sa2.explore()
routes.explore(m = m)
stops.explore(m = m)

We'll look at what's in the datasets a bit more closely as we get into the various GIS methods available.
