==> 12-making-a-script.ipynb <==
# Making a notebook into a script
It's easy to make a notebook into a script.

First write some code

def greeting(name = "world"):
    print(f"Hello {name}")
    
if __name__ == "__main__":
    greeting()

You save it to a script in VSCode with `<cmd><shift><p>` then find the export to script option. I saved this one to `simple-script.py`. You can run it from the command line with 

    python3 simple-script.py

That's pretty much it for a simple script. The only quirk is the `if __name__ == "__main__":` clause. This is there to prevent the running parts of the code (which might overwrite data say) from running if you invoke the code by importing it into another file, when you probably only want the functions contained in the file and not the side-effects of running it.

Many developers recommend against using notebooks for code development. I can see the problems, but for simple scripts that are largely linear (like many data processing pipelines) I think it is fine to figure out the details interactively in a notebook and then deploy as a script. 

## Passing arguments into a script
The main additional feature desirable in a script is how to pass parameters into the script at run time. This is also not difficult. Here's a version of the code for the above hello world that can take in a command line parameter. It uses `sys.argv`. Consult the documentation for more about this.

```python
import sys

def greeting(name = "world"):
    print(f"Hello {name}")

if __name__ == "__main__":
    if len(sys.argv) > 1:
        greeting(sys.argv[1])
    else:
        greeting()
```

`sys.argv` is a list containing all the values in the command line. Item 0 is the script name, item 1 the next one, item 2 the next one, and so on. This approach makes it simple to develop scripts that you can point at a specified file.

## More complicated arguments and help
The [`argparse`](https://docs.python.org/3/library/argparse.html#module-argparse) module provides options to make much more complex command line options which may be desirable for data processing pipeline scripts.

```python
import argparse
parser = argparse.ArgumentParser()
parser.add_argument(
    "-n", "--name", help = "name to greet")
parser.parse_args()
args = parser.parse_args()

def greeting(name = "world"):
    print(f"Hello {name}")

if __name__ == "__main__":
    if args.name is None:
        greeting()
    else:
        greeting(args.name)
```

The [tutorial materials here](https://docs.python.org/3/howto/argparse.html) provide an easily following guide to developing complex command line options.
